+++ blueOceanField/domain/market.py
@@ -49,9 +49,10 @@ class ExchangePlace:
         if not self.name:
             raise ValueError("ExchangePlace name cannot be empty")
 
-    @classmethod
-    def backtest(cls):
-        return cls("backtest")
+    @staticmethod
+    @property
+    def BACKTEST():
+        return ExchangePlace("backtest")
 
 @dataclass(frozen=True)
 class Ohlcv:
