+++ blueOceanField/presentation/grpc.py
@@ -1,18 +1,8 @@
-from injector import Module, inject, singleton
 from blueOceanField.application.converter import GrpcConverter
 import blueOceanField.application.generated as proto
 import blueOceanField.application.usecase as usecase
-from blueOceanField.domain.market import IExchange, IOhlcvRepository
-
-
-class GrpcModule(Module):
-    def configure(self, binder):
-        binder.bind(
-            proto.FeatureProcessServiceServicer,
-            to=FeatureProcessHandler,
-            scope=singleton,
-        )
-        binder.bind(proto.MarketServiceServicer, to=MarketHandler, scope=singleton)
+from blueOceanField.domain.market import IExchange
+from blueOceanField.application.container.context import AppContext
 
 
 class FeatureProcessHandler(proto.FeatureProcessServiceServicer):
@@ -27,8 +17,6 @@ class MarketHandler(proto.MarketServiceServicer):
         return proto.GetExchangesResponse(exchangePlaces=places)
 
     async def GetSymbols(self, request: proto.GetSymbolsRequest, context):
-        from blueOceanField.presentation.context import AppContext
-
         target_exchange = GrpcConverter.from_grpc(request.exchange)
         exchange_context = AppContext.get_or_create_exchange_context(target_exchange)
         symbols = [
